changing some things now works as midi fx in logic pro x

+++ first-plugin/Source/PluginEditor.cpp
@@ -35,5 +35,5 @@ void HelloWorldPluginAudioProcessorEditor::resized() {
 }
 
 void HelloWorldPluginAudioProcessorEditor::sliderValueChanged(juce::Slider* slider) {
-    processorRef.noteOnVel = midiVolume.getValue();
+    processorRef.noteOnVel = (float)midiVolume.getValue();
 }
